feat(lib: libalgo): add AVL tree implementation
I plan to use an AVL to store VMAs inside the VMM.

This implementation is not responsible for allocating the node's
structure, as it is intended to be used with intrusive C structs.
The caller has to pass the address of the field of a bigger structure,
reprensenting the node's data.

This is inpired by linux's intrusive linked list design:
* https://www.data-structures-in-practice.com/intrusive-linked-lists

I used this opportunity to create a separate "libalgo" library, which
will contain the implementation of all the datastructures we need and
use throughout this project.

Signed-off by: Léo DUBOIN <leo@duboin.com>

+++ kernel/include/utils/types.h
@@ -16,6 +16,6 @@ typedef int64_t s64;
 typedef float f32;  //< 32b floating point value
 typedef double f64; //< 64b floating point value
 
-typedef int32_t ssize_t;
+typedef long int ssize_t;
 
 #endif /* UTILS_TYPES_H */
